requirements.txt
Flask==2.3.2
flask_cors==3.0.10
pymysql==1.0.3
PyYAML==6.0.1